Port to .NET 5

+++ VirtualGameMode/Utilities/NotifyIconPosition.cs
@@ -11,12 +11,12 @@ namespace VirtualGameMode.Utilities
         {
             // obtain private field "id"
             // ReSharper disable once PossibleNullReferenceException
-            var id = (int)icon.GetType().GetField("id", BindingFlags.NonPublic | BindingFlags.Instance).GetValue(icon);
+            var id = (uint)icon.GetType().GetField("id", BindingFlags.NonPublic | BindingFlags.Instance).GetValue(icon);
             // ReSharper disable once PossibleNullReferenceException
             var hwnd = (NativeWindow) icon.GetType()
                 .GetField("window", BindingFlags.NonPublic | BindingFlags.Instance).GetValue(icon);
 
-            Native.NOTIFYICONIDENTIFIER identifier = new Native.NOTIFYICONIDENTIFIER() {cbSize = Marshal.SizeOf(typeof(Native.NOTIFYICONIDENTIFIER)),hWnd = hwnd.Handle, uID = id};
+            Native.NOTIFYICONIDENTIFIER identifier = new Native.NOTIFYICONIDENTIFIER() {cbSize = Marshal.SizeOf(typeof(Native.NOTIFYICONIDENTIFIER)),hWnd = hwnd.Handle, uID = (int)id};
 
             Native.Shell_NotifyIconGetRect(ref identifier, out var rect);
             return Rectangle.FromLTRB(rect.Left, rect.Top, rect.Right, rect.Bottom);
